dataimporter/requirements.txt
pandas~=2.2.3
psycopg2-binary